Added animate to validate card on tables and categories

+++ app/sections/table/Card.tsx
@@ -4,6 +4,9 @@ import { Icon } from "@iconify/react";
 import { toast } from "sonner";
 import { useTableStore } from "~/store/tableStore";
 import { Table as TableType } from "types";
+import { useLoadingStore } from "~/store/loadingStore";
+import { motion } from "framer-motion";
+import { useAnimateFeedbackCard } from "~/hooks/useAnimateFeedbackCard";
 
 interface TableCardProps {
   table: TableType & { is_persistent: boolean };
@@ -19,53 +22,88 @@ const TableCard: React.FC<TableCardProps> = ({ table }) => {
   const [isEditing, setIsEditing] = React.useState(!table.is_persistent);
   const [number, setNumber] = React.useState(table.number);
   const inputRef = useRef<HTMLInputElement>(null);
+  const withLoading = useLoadingStore((state) => state.withLoading);
+  const { controls, animate } = useAnimateFeedbackCard();
 
   useEffect(() => {
     if (isEditing && inputRef.current) {
       inputRef.current.focus(); // Autofocus al input
+    }
+
+    if (!isEditing && number.trim() === "") {
+      setNumber(table.number);
     }
   }, [isEditing]);
 
   useEffect(() => {
     if (fetcher.data) {
       if (fetcher.data.error) {
-        toast.warning(fetcher.data.error);
+        toast.error(fetcher.data.error);
+        animate("error");
         inputRef.current?.focus(); // Autofocus al input
         return;
       }
+
       if (fetcher.data.table) {
         updateTable({ is_persistent: true, ...fetcher.data.table }, table.id);
         toast.success(fetcher.data.message);
+        animate("success");
         setIsEditing(false);
         return;
       }
+
+      if (fetcher.data.message) {
+        toast.success(fetcher.data.message); // Mensaje del servidor
+        animate("success");
+        deleteTable(table.id);
+        return;
+      }
     }
-  }, [fetcher.data, updateTable]);
+  }, [fetcher.data]);
+
+  const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
+    if (event.key === "Escape") {
+      if (!table.is_persistent) {
+        deleteTable(table.id);
+      }
+      setIsEditing(false);
+    }
+  };
 
   const handleSave = async () => {
     if (number.trim() === "") {
-      toast.error("El número de la mesa no puede estar vacío");
+      toast.warning("El número de la mesa no puede estar vacío");
+      animate("error");
       return;
     }
 
-    const action = !table.is_persistent
-      ? "/admin/tables/create"
-      : `/admin/tables/${table.id}/edit`;
-    await fetcher.submit({ number }, { method: "post", action });
+    //Verificar si el número ha cambiado
+    if (table.number !== number) {
+      const action = !table.is_persistent
+        ? "/admin/tables/create"
+        : `/admin/tables/${table.id}/edit`;
+
+      await withLoading(async () => {
+        await fetcher.submit({ number }, { method: "post", action });
+      });
+    } else {
+      setIsEditing(false);
+    }
   };
 
   const handleDelete = async () => {
-    await fetcher.submit(null, {
-      method: "delete",
-      action: `/admin/tables/${table.id}/destroy`,
+    await withLoading(async () => {
+      await fetcher.submit(null, {
+        method: "delete",
+        action: `/admin/tables/${table.id}/destroy`,
+      });
     });
-    deleteTable(table.id);
   };
 
   return (
-    <div
+    <motion.div
+      animate={controls}
       className="p-4 border rounded shadow hover:shadow-md transition select-none"
-      onDoubleClick={() => setIsEditing(true)}
     >
       <div className="flex items-center justify-between">
         {isEditing ? (
@@ -80,11 +118,15 @@ const TableCard: React.FC<TableCardProps> = ({ table }) => {
               type="text"
               value={number}
               onChange={(e) => setNumber(e.target.value)}
+              onKeyDown={handleKeyDown}
               className="border px-2 py-1 rounded w-full"
             />
           </form>
         ) : (
-          <span className="text-lg font-medium text-primary-900">
+          <span
+            className="text-lg font-medium text-primary-900"
+            onDoubleClick={() => setIsEditing(true)}
+          >
             {`Mesa #${table.number || "Nueva Mesa"}${
               table.is_persistent ? "" : "*"
             }`}
@@ -109,7 +151,7 @@ const TableCard: React.FC<TableCardProps> = ({ table }) => {
               ) : (
                 <button
                   className="text-red-500 hover:text-red-600"
-                  onClick={handleDelete}
+                  onClick={() => deleteTable(table.id)}
                 >
                   <Icon className="h-5 w-5" icon="tabler:trash" />
                 </button>
@@ -133,7 +175,7 @@ const TableCard: React.FC<TableCardProps> = ({ table }) => {
           )}
         </div>
       </div>
-    </div>
+    </motion.div>
   );
 };
